app/channel,config,memory: extract support modules
