+++ crates/app/src/channel/registry.rs
@@ -58,6 +58,9 @@ mod plugin_bridge;
 #[path = "registry_surface.rs"]
 mod surface_support;
 
+#[path = "registry_status.rs"]
+mod status_support;
+
 #[cfg(test)]
 #[path = "registry_plugin_bridge_tests.rs"]
 mod plugin_bridge_tests;
@@ -93,6 +96,18 @@ pub use plugin_bridge::{
 };
 use plugin_bridge::{
     channel_surface_plugin_bridge_discovery_by_id, plugin_bridge_contract_from_descriptor,
+};
+use status_support::{
+    attach_runtime, build_invalid_dingtalk_snapshot, build_invalid_discord_snapshot,
+    build_invalid_email_snapshot, build_invalid_feishu_snapshot,
+    build_invalid_google_chat_snapshot, build_invalid_imessage_snapshot,
+    build_invalid_irc_snapshot, build_invalid_line_snapshot, build_invalid_matrix_snapshot,
+    build_invalid_mattermost_snapshot, build_invalid_nextcloud_talk_snapshot,
+    build_invalid_signal_snapshot, build_invalid_slack_snapshot,
+    build_invalid_synology_chat_snapshot, build_invalid_teams_snapshot,
+    build_invalid_telegram_snapshot, build_invalid_webhook_snapshot, build_invalid_wecom_snapshot,
+    build_invalid_whatsapp_snapshot, disabled_operation, misconfigured_operation, ready_operation,
+    unsupported_operation,
 };
 use surface_support::build_channel_surfaces;
 
@@ -5688,1228 +5703,6 @@ fn build_wecom_snapshot_for_account(
         reserved_runtime_fields: Vec::new(),
         operations: vec![send_operation, serve_operation],
     }
-}
-
-fn build_invalid_telegram_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            TELEGRAM_SEND_OPERATION,
-            "binary built without feature `channel-telegram`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(TELEGRAM_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            TELEGRAM_SERVE_OPERATION,
-            "binary built without feature `channel-telegram`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(TELEGRAM_SERVE_OPERATION, vec![error.clone()])
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_feishu_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            FEISHU_SEND_OPERATION,
-            "binary built without feature `channel-feishu`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(FEISHU_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            FEISHU_SERVE_OPERATION,
-            "binary built without feature `channel-feishu`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(FEISHU_SERVE_OPERATION, vec![error.clone()])
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_matrix_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            MATRIX_SEND_OPERATION,
-            "binary built without feature `channel-matrix`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(MATRIX_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            MATRIX_SERVE_OPERATION,
-            "binary built without feature `channel-matrix`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(MATRIX_SERVE_OPERATION, vec![error.clone()])
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_wecom_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            WECOM_SEND_OPERATION,
-            "binary built without feature `channel-wecom`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(WECOM_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            WECOM_SERVE_OPERATION,
-            "binary built without feature `channel-wecom`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(WECOM_SERVE_OPERATION, vec![error.clone()])
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_discord_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            DISCORD_SEND_OPERATION,
-            "binary built without feature `channel-discord`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(DISCORD_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            DISCORD_SERVE_OPERATION,
-            "binary built without feature `channel-discord`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            DISCORD_SERVE_OPERATION,
-            "discord serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_slack_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            SLACK_SEND_OPERATION,
-            "binary built without feature `channel-slack`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(SLACK_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            SLACK_SERVE_OPERATION,
-            "binary built without feature `channel-slack`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            SLACK_SERVE_OPERATION,
-            "slack serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_line_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-    runtime_dir: &Path,
-    now_ms: u64,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            LINE_SEND_OPERATION,
-            "binary built without feature `channel-line`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(LINE_SEND_OPERATION, vec![error.clone()])
-    };
-    let send_operation = attach_runtime(
-        ChannelPlatform::Line,
-        LINE_SEND_OPERATION,
-        send_operation,
-        configured_account_id,
-        configured_account_id,
-        runtime_dir,
-        now_ms,
-    );
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            LINE_SERVE_OPERATION,
-            "binary built without feature `channel-line`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(LINE_SERVE_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = attach_runtime(
-        ChannelPlatform::Line,
-        LINE_SERVE_OPERATION,
-        serve_operation,
-        configured_account_id,
-        configured_account_id,
-        runtime_dir,
-        now_ms,
-    );
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_dingtalk_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            DINGTALK_SEND_OPERATION,
-            "binary built without feature `channel-dingtalk`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(DINGTALK_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            DINGTALK_SERVE_OPERATION,
-            "binary built without feature `channel-dingtalk`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            DINGTALK_SERVE_OPERATION,
-            "dingtalk custom robot surface is outbound-only".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_whatsapp_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            WHATSAPP_SEND_OPERATION,
-            "binary built without feature `channel-whatsapp`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(WHATSAPP_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            WHATSAPP_SERVE_OPERATION,
-            "binary built without feature `channel-whatsapp`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(WHATSAPP_SERVE_OPERATION, vec![error.clone()])
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_email_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            EMAIL_SEND_OPERATION,
-            "binary built without feature `channel-email`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(EMAIL_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            EMAIL_SERVE_OPERATION,
-            "binary built without feature `channel-email`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            EMAIL_SERVE_OPERATION,
-            "email IMAP reply-loop serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_webhook_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-    runtime_dir: &Path,
-    now_ms: u64,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            WEBHOOK_SEND_OPERATION,
-            "binary built without feature `channel-webhook`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(WEBHOOK_SEND_OPERATION, vec![error.clone()])
-    };
-    let send_operation = attach_runtime(
-        ChannelPlatform::Webhook,
-        WEBHOOK_SEND_OPERATION,
-        send_operation,
-        configured_account_id,
-        configured_account_id,
-        runtime_dir,
-        now_ms,
-    );
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            WEBHOOK_SERVE_OPERATION,
-            "binary built without feature `channel-webhook`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(WEBHOOK_SERVE_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = attach_runtime(
-        ChannelPlatform::Webhook,
-        WEBHOOK_SERVE_OPERATION,
-        serve_operation,
-        configured_account_id,
-        configured_account_id,
-        runtime_dir,
-        now_ms,
-    );
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_google_chat_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            GOOGLE_CHAT_SEND_OPERATION,
-            "binary built without feature `channel-google-chat`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(GOOGLE_CHAT_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            GOOGLE_CHAT_SERVE_OPERATION,
-            "binary built without feature `channel-google-chat`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            GOOGLE_CHAT_SERVE_OPERATION,
-            "google chat incoming webhook surface is outbound-only".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_signal_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            SIGNAL_SEND_OPERATION,
-            "binary built without feature `channel-signal`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(SIGNAL_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            SIGNAL_SERVE_OPERATION,
-            "binary built without feature `channel-signal`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            SIGNAL_SERVE_OPERATION,
-            "signal serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_irc_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            IRC_SEND_OPERATION,
-            "binary built without feature `channel-irc`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(IRC_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            IRC_SERVE_OPERATION,
-            "binary built without feature `channel-irc`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            IRC_SERVE_OPERATION,
-            "irc relay-loop serve is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_teams_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            TEAMS_SEND_OPERATION,
-            "binary built without feature `channel-teams`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(TEAMS_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            TEAMS_SERVE_OPERATION,
-            "binary built without feature `channel-teams`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            TEAMS_SERVE_OPERATION,
-            "microsoft teams incoming webhook surface is outbound-only today".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_imessage_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            IMESSAGE_SEND_OPERATION,
-            "binary built without feature `channel-imessage`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(IMESSAGE_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            IMESSAGE_SERVE_OPERATION,
-            "binary built without feature `channel-imessage`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            IMESSAGE_SERVE_OPERATION,
-            "imessage bridge sync runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_mattermost_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            MATTERMOST_SEND_OPERATION,
-            "binary built without feature `channel-mattermost`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(MATTERMOST_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            MATTERMOST_SERVE_OPERATION,
-            "binary built without feature `channel-mattermost`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            MATTERMOST_SERVE_OPERATION,
-            "mattermost serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_nextcloud_talk_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            NEXTCLOUD_TALK_SEND_OPERATION,
-            "binary built without feature `channel-nextcloud-talk`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(NEXTCLOUD_TALK_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            NEXTCLOUD_TALK_SERVE_OPERATION,
-            "binary built without feature `channel-nextcloud-talk`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            NEXTCLOUD_TALK_SERVE_OPERATION,
-            "nextcloud talk bot callback serve is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_invalid_synology_chat_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            SYNOLOGY_CHAT_SEND_OPERATION,
-            "binary built without feature `channel-synology-chat`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(SYNOLOGY_CHAT_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            SYNOLOGY_CHAT_SERVE_OPERATION,
-            "binary built without feature `channel-synology-chat`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            SYNOLOGY_CHAT_SERVE_OPERATION,
-            "synology chat outgoing webhook serve is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        reserved_runtime_fields: Vec::new(),
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn ready_operation(operation: ChannelCatalogOperation) -> ChannelOperationStatus {
-    ChannelOperationStatus {
-        id: operation.id,
-        label: operation.label,
-        command: operation.command,
-        health: ChannelOperationHealth::Ready,
-        detail: "ready".to_owned(),
-        issues: Vec::new(),
-        runtime: None,
-    }
-}
-
-fn disabled_operation(
-    operation: ChannelCatalogOperation,
-    detail: String,
-) -> ChannelOperationStatus {
-    ChannelOperationStatus {
-        id: operation.id,
-        label: operation.label,
-        command: operation.command,
-        health: ChannelOperationHealth::Disabled,
-        detail,
-        issues: Vec::new(),
-        runtime: None,
-    }
-}
-
-fn unsupported_operation(
-    operation: ChannelCatalogOperation,
-    detail: String,
-) -> ChannelOperationStatus {
-    ChannelOperationStatus {
-        id: operation.id,
-        label: operation.label,
-        command: operation.command,
-        health: ChannelOperationHealth::Unsupported,
-        detail: detail.clone(),
-        issues: vec![detail],
-        runtime: None,
-    }
-}
-
-fn misconfigured_operation(
-    operation: ChannelCatalogOperation,
-    issues: Vec<String>,
-) -> ChannelOperationStatus {
-    ChannelOperationStatus {
-        id: operation.id,
-        label: operation.label,
-        command: operation.command,
-        health: ChannelOperationHealth::Misconfigured,
-        detail: issues.join("; "),
-        issues,
-        runtime: None,
-    }
-}
-
-fn attach_runtime(
-    platform: ChannelPlatform,
-    operation: ChannelCatalogOperation,
-    mut status: ChannelOperationStatus,
-    account_id: &str,
-    account_label: &str,
-    runtime_dir: &Path,
-    now_ms: u64,
-) -> ChannelOperationStatus {
-    if operation.tracks_runtime {
-        status.runtime = state::load_channel_operation_runtime_for_account_from_dir(
-            runtime_dir,
-            platform,
-            operation.id,
-            account_id,
-            now_ms,
-        )
-        .map(|mut runtime| {
-            if runtime.account_id.is_none() {
-                runtime.account_id = Some(account_id.to_owned());
-            }
-            if runtime.account_label.is_none() {
-                runtime.account_label = Some(account_label.to_owned());
-            }
-            runtime
-        })
-        .or(Some(ChannelOperationRuntime {
-            running: false,
-            stale: false,
-            busy: false,
-            active_runs: 0,
-            last_run_activity_at: None,
-            last_heartbeat_at: None,
-            pid: None,
-            account_id: Some(account_id.to_owned()),
-            account_label: Some(account_label.to_owned()),
-            instance_count: 0,
-            running_instances: 0,
-            stale_instances: 0,
-        }));
-    }
-    status
 }
 
 fn now_ms() -> u64 {
