+++ crates/app/src/config/channels.rs
@@ -4,10 +4,6 @@ use loong_contracts::SecretRef;
 use serde::{Deserialize, Serialize};
 
 use crate::CliResult;
-use crate::prompt::{
-    DEFAULT_PROMPT_PACK_ID, PromptPersonality, PromptRenderInput, render_default_system_prompt,
-    render_system_prompt,
-};
 
 use super::irc::{
     IRC_NICKNAME_ENV, IRC_PASSWORD_ENV, IRC_SERVER_ENV, default_irc_nickname_env,
@@ -32,14 +28,24 @@ mod twitch;
 
 #[allow(unused_imports)]
 pub use self::twitch::{ResolvedTwitchChannelConfig, TwitchAccountConfig, TwitchChannelConfig};
+#[path = "channels_account_resolution.rs"]
+mod account_resolution;
 #[path = "channels_defaults.rs"]
 mod defaults;
+#[path = "channels_email_support.rs"]
+mod email_support;
+#[path = "channels_shared_types.rs"]
+mod shared_types;
 
 use self::defaults::*;
+pub(crate) use self::email_support::parse_email_smtp_endpoint;
+use self::email_support::*;
+use account_resolution::*;
 #[path = "channels_validation.rs"]
 mod validation_support;
 
 use self::validation_support::*;
+pub(crate) use account_resolution::normalize_channel_account_id;
 #[allow(unused_imports)]
 pub use bridge::{
     OnebotAccountConfig, OnebotChannelConfig, QqbotAccountConfig, QqbotChannelConfig,
@@ -48,128 +54,33 @@ pub use bridge::{
 };
 pub use nostr_impl::{NostrAccountConfig, NostrChannelConfig, ResolvedNostrChannelConfig};
 pub(crate) use nostr_impl::{parse_nostr_private_key_hex, parse_nostr_public_key_hex};
+pub use shared_types::{
+    ChannelAccountIdentity, ChannelAccountIdentitySource, ChannelAcpConfig,
+    ChannelDefaultAccountSelection, ChannelDefaultAccountSelectionSource,
+    ChannelResolvedAccountRoute, CliChannelConfig, FeishuDomain, TelegramStreamingMode,
+    WebhookPayloadFormat,
+};
+pub(crate) use shared_types::{
+    DINGTALK_SECRET_ENV, DINGTALK_WEBHOOK_URL_ENV, DISCORD_APPLICATION_ID_ENV,
+    DISCORD_BOT_TOKEN_ENV, EMAIL_IMAP_PASSWORD_ENV, EMAIL_IMAP_USERNAME_ENV,
+    EMAIL_SMTP_PASSWORD_ENV, EMAIL_SMTP_USERNAME_ENV, EmailSmtpEndpoint, FEISHU_APP_ID_ENV,
+    FEISHU_APP_SECRET_ENV, FEISHU_ENCRYPT_KEY_ENV, FEISHU_VERIFICATION_TOKEN_ENV,
+    GOOGLE_CHAT_WEBHOOK_URL_ENV, IMESSAGE_BRIDGE_TOKEN_ENV, IMESSAGE_BRIDGE_URL_ENV,
+    LINE_CHANNEL_ACCESS_TOKEN_ENV, LINE_CHANNEL_SECRET_ENV, MATRIX_ACCESS_TOKEN_ENV,
+    MATTERMOST_BOT_TOKEN_ENV, MATTERMOST_SERVER_URL_ENV, NEXTCLOUD_TALK_SERVER_URL_ENV,
+    NEXTCLOUD_TALK_SHARED_SECRET_ENV, NOSTR_PRIVATE_KEY_ENV, NOSTR_RELAY_URLS_ENV,
+    ONEBOT_ACCESS_TOKEN_ENV, ONEBOT_WEBSOCKET_URL_ENV, QQBOT_APP_ID_ENV, QQBOT_CLIENT_SECRET_ENV,
+    SIGNAL_ACCOUNT_ENV, SIGNAL_SERVICE_URL_ENV, SLACK_BOT_TOKEN_ENV,
+    SYNOLOGY_CHAT_INCOMING_URL_ENV, SYNOLOGY_CHAT_TOKEN_ENV, TEAMS_APP_ID_ENV,
+    TEAMS_APP_PASSWORD_ENV, TEAMS_TENANT_ID_ENV, TEAMS_WEBHOOK_URL_ENV, TELEGRAM_BOT_TOKEN_ENV,
+    TLON_CODE_ENV, TLON_SHIP_ENV, TLON_URL_ENV, TWITCH_ACCESS_TOKEN_ENV, WEBHOOK_AUTH_TOKEN_ENV,
+    WEBHOOK_ENDPOINT_URL_ENV, WEBHOOK_SIGNING_SECRET_ENV, WECOM_BOT_ID_ENV, WECOM_SECRET_ENV,
+    WEIXIN_BRIDGE_ACCESS_TOKEN_ENV, WEIXIN_BRIDGE_URL_ENV, WHATSAPP_ACCESS_TOKEN_ENV,
+    WHATSAPP_APP_SECRET_ENV, WHATSAPP_PHONE_NUMBER_ID_ENV, WHATSAPP_VERIFY_TOKEN_ENV,
+};
 use signal_impl::{
     default_signal_account_env, default_signal_service_url, default_signal_service_url_env,
 };
-
-pub(crate) const TELEGRAM_BOT_TOKEN_ENV: &str = "TELEGRAM_BOT_TOKEN";
-pub(crate) const DISCORD_BOT_TOKEN_ENV: &str = "DISCORD_BOT_TOKEN";
-pub(crate) const DISCORD_APPLICATION_ID_ENV: &str = "DISCORD_APPLICATION_ID";
-pub(crate) const DINGTALK_WEBHOOK_URL_ENV: &str = "DINGTALK_WEBHOOK_URL";
-pub(crate) const DINGTALK_SECRET_ENV: &str = "DINGTALK_SECRET";
-pub(crate) const EMAIL_SMTP_USERNAME_ENV: &str = "EMAIL_SMTP_USERNAME";
-pub(crate) const EMAIL_SMTP_PASSWORD_ENV: &str = "EMAIL_SMTP_PASSWORD";
-pub(crate) const EMAIL_IMAP_USERNAME_ENV: &str = "EMAIL_IMAP_USERNAME";
-pub(crate) const EMAIL_IMAP_PASSWORD_ENV: &str = "EMAIL_IMAP_PASSWORD";
-pub(crate) const FEISHU_APP_ID_ENV: &str = "FEISHU_APP_ID";
-pub(crate) const FEISHU_APP_SECRET_ENV: &str = "FEISHU_APP_SECRET";
-pub(crate) const FEISHU_VERIFICATION_TOKEN_ENV: &str = "FEISHU_VERIFICATION_TOKEN";
-pub(crate) const FEISHU_ENCRYPT_KEY_ENV: &str = "FEISHU_ENCRYPT_KEY";
-pub(crate) const GOOGLE_CHAT_WEBHOOK_URL_ENV: &str = "GOOGLE_CHAT_WEBHOOK_URL";
-pub(crate) const LINE_CHANNEL_ACCESS_TOKEN_ENV: &str = "LINE_CHANNEL_ACCESS_TOKEN";
-pub(crate) const LINE_CHANNEL_SECRET_ENV: &str = "LINE_CHANNEL_SECRET";
-pub(crate) const MATRIX_ACCESS_TOKEN_ENV: &str = "MATRIX_ACCESS_TOKEN";
-pub(crate) const MATTERMOST_SERVER_URL_ENV: &str = "MATTERMOST_SERVER_URL";
-pub(crate) const MATTERMOST_BOT_TOKEN_ENV: &str = "MATTERMOST_BOT_TOKEN";
-pub(crate) const NEXTCLOUD_TALK_SERVER_URL_ENV: &str = "NEXTCLOUD_TALK_SERVER_URL";
-pub(crate) const NEXTCLOUD_TALK_SHARED_SECRET_ENV: &str = "NEXTCLOUD_TALK_SHARED_SECRET";
-pub(crate) const SYNOLOGY_CHAT_TOKEN_ENV: &str = "SYNOLOGY_CHAT_TOKEN";
-pub(crate) const SYNOLOGY_CHAT_INCOMING_URL_ENV: &str = "SYNOLOGY_CHAT_INCOMING_URL";
-pub(crate) const SIGNAL_SERVICE_URL_ENV: &str = "SIGNAL_SERVICE_URL";
-pub(crate) const SIGNAL_ACCOUNT_ENV: &str = "SIGNAL_ACCOUNT";
-pub(crate) const TWITCH_ACCESS_TOKEN_ENV: &str = "TWITCH_ACCESS_TOKEN";
-pub(crate) const SLACK_BOT_TOKEN_ENV: &str = "SLACK_BOT_TOKEN";
-pub(crate) const TEAMS_APP_ID_ENV: &str = "TEAMS_APP_ID";
-pub(crate) const TEAMS_APP_PASSWORD_ENV: &str = "TEAMS_APP_PASSWORD";
-pub(crate) const TEAMS_TENANT_ID_ENV: &str = "TEAMS_TENANT_ID";
-pub(crate) const TEAMS_WEBHOOK_URL_ENV: &str = "TEAMS_WEBHOOK_URL";
-pub(crate) const IMESSAGE_BRIDGE_URL_ENV: &str = "IMESSAGE_BRIDGE_URL";
-pub(crate) const IMESSAGE_BRIDGE_TOKEN_ENV: &str = "IMESSAGE_BRIDGE_TOKEN";
-pub(crate) const NOSTR_RELAY_URLS_ENV: &str = "NOSTR_RELAY_URLS";
-pub(crate) const NOSTR_PRIVATE_KEY_ENV: &str = "NOSTR_PRIVATE_KEY";
-pub(crate) const TLON_SHIP_ENV: &str = "TLON_SHIP";
-pub(crate) const TLON_URL_ENV: &str = "TLON_URL";
-pub(crate) const TLON_CODE_ENV: &str = "TLON_CODE";
-pub(crate) const WEIXIN_BRIDGE_URL_ENV: &str = "WEIXIN_BRIDGE_URL";
-pub(crate) const WEIXIN_BRIDGE_ACCESS_TOKEN_ENV: &str = "WEIXIN_BRIDGE_ACCESS_TOKEN";
-pub(crate) const QQBOT_APP_ID_ENV: &str = "QQBOT_APP_ID";
-pub(crate) const QQBOT_CLIENT_SECRET_ENV: &str = "QQBOT_CLIENT_SECRET";
-pub(crate) const ONEBOT_WEBSOCKET_URL_ENV: &str = "ONEBOT_WEBSOCKET_URL";
-pub(crate) const ONEBOT_ACCESS_TOKEN_ENV: &str = "ONEBOT_ACCESS_TOKEN";
-pub(crate) const WHATSAPP_ACCESS_TOKEN_ENV: &str = "WHATSAPP_ACCESS_TOKEN";
-pub(crate) const WHATSAPP_PHONE_NUMBER_ID_ENV: &str = "WHATSAPP_PHONE_NUMBER_ID";
-pub(crate) const WHATSAPP_VERIFY_TOKEN_ENV: &str = "WHATSAPP_VERIFY_TOKEN";
-pub(crate) const WHATSAPP_APP_SECRET_ENV: &str = "WHATSAPP_APP_SECRET";
-pub(crate) const WECOM_BOT_ID_ENV: &str = "WECOM_BOT_ID";
-pub(crate) const WECOM_SECRET_ENV: &str = "WECOM_SECRET";
-pub(crate) const WEBHOOK_ENDPOINT_URL_ENV: &str = "WEBHOOK_ENDPOINT_URL";
-pub(crate) const WEBHOOK_AUTH_TOKEN_ENV: &str = "WEBHOOK_AUTH_TOKEN";
-pub(crate) const WEBHOOK_SIGNING_SECRET_ENV: &str = "WEBHOOK_SIGNING_SECRET";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum TelegramStreamingMode {
-    #[default]
-    Off,
-    Draft,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum WebhookPayloadFormat {
-    #[default]
-    JsonText,
-    PlainText,
-}
-
-impl WebhookPayloadFormat {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::JsonText => "json_text",
-            Self::PlainText => "plain_text",
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum EmailSmtpEndpoint {
-    RelayHost(String),
-    ConnectionUrl(String),
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct CliChannelConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default = "default_system_prompt")]
-    pub system_prompt: String,
-    #[serde(default = "default_prompt_pack_id")]
-    pub prompt_pack_id: Option<String>,
-    #[serde(default = "default_prompt_personality")]
-    pub personality: Option<PromptPersonality>,
-    #[serde(default)]
-    pub system_prompt_addendum: Option<String>,
-    #[serde(default = "default_exit_commands")]
-    pub exit_commands: Vec<String>,
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ChannelAcpConfig {
-    #[serde(default)]
-    pub bootstrap_mcp_servers: Vec<String>,
-    #[serde(default)]
-    pub working_directory: Option<String>,
-}
-
-impl ChannelAcpConfig {
-    pub fn resolved_working_directory(&self) -> Option<std::path::PathBuf> {
-        self.working_directory
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(std::path::PathBuf::from)
-    }
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct TelegramChannelConfig {
@@ -201,91 +112,6 @@ pub struct TelegramChannelConfig {
     pub ack_reactions: bool,
     #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
     pub accounts: BTreeMap<String, TelegramAccountConfig>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum FeishuDomain {
-    #[default]
-    Feishu,
-    Lark,
-}
-
-impl FeishuDomain {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Feishu => "feishu",
-            Self::Lark => "lark",
-        }
-    }
-
-    pub fn default_base_url(self) -> &'static str {
-        match self {
-            Self::Feishu => "https://open.feishu.cn",
-            Self::Lark => "https://open.larksuite.com",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ChannelAccountIdentitySource {
-    Configured,
-    DerivedCredential,
-    Default,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ChannelDefaultAccountSelectionSource {
-    ExplicitDefault,
-    MappedDefault,
-    Fallback,
-    RuntimeIdentity,
-}
-
-impl ChannelDefaultAccountSelectionSource {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::ExplicitDefault => "explicit_default",
-            Self::MappedDefault => "mapped_default",
-            Self::Fallback => "fallback",
-            Self::RuntimeIdentity => "runtime_identity",
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-pub struct ChannelAccountIdentity {
-    pub id: String,
-    pub label: String,
-    pub source: ChannelAccountIdentitySource,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-pub struct ChannelDefaultAccountSelection {
-    pub id: String,
-    pub source: ChannelDefaultAccountSelectionSource,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-pub struct ChannelResolvedAccountRoute {
-    pub requested_account_id: Option<String>,
-    pub configured_account_count: usize,
-    pub selected_configured_account_id: String,
-    pub default_account_source: ChannelDefaultAccountSelectionSource,
-}
-
-impl ChannelResolvedAccountRoute {
-    pub fn selected_by_default(&self) -> bool {
-        self.requested_account_id.is_none()
-    }
-
-    pub fn uses_implicit_fallback_default(&self) -> bool {
-        self.selected_by_default()
-            && self.configured_account_count > 1
-            && self.default_account_source == ChannelDefaultAccountSelectionSource::Fallback
-    }
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
@@ -1945,62 +1771,6 @@ pub struct TlonChannelConfig {
     pub code_env: Option<String>,
     #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
     pub accounts: BTreeMap<String, TlonAccountConfig>,
-}
-
-impl Default for CliChannelConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            system_prompt: default_system_prompt(),
-            prompt_pack_id: default_prompt_pack_id(),
-            personality: default_prompt_personality(),
-            system_prompt_addendum: None,
-            exit_commands: default_exit_commands(),
-        }
-    }
-}
-
-impl CliChannelConfig {
-    pub fn uses_native_prompt_pack(&self) -> bool {
-        self.prompt_pack_id().is_some_and(|value| !value.is_empty())
-    }
-
-    pub fn prompt_pack_id(&self) -> Option<&str> {
-        self.prompt_pack_id
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-    }
-
-    pub fn resolved_personality(&self) -> PromptPersonality {
-        self.personality.unwrap_or_default()
-    }
-
-    pub fn rendered_native_system_prompt(&self) -> String {
-        render_system_prompt(PromptRenderInput {
-            personality: self.resolved_personality(),
-            addendum: self.system_prompt_addendum.clone(),
-        })
-    }
-
-    pub fn resolved_system_prompt(&self) -> String {
-        if self.uses_native_prompt_pack() {
-            return self.rendered_native_system_prompt();
-        }
-
-        let inline = self.system_prompt.trim();
-        if !inline.is_empty() {
-            return inline.to_owned();
-        }
-
-        render_default_system_prompt()
-    }
-
-    pub fn refresh_native_system_prompt(&mut self) {
-        if self.uses_native_prompt_pack() {
-            self.system_prompt = self.rendered_native_system_prompt();
-        }
-    }
 }
 
 impl Default for TelegramChannelConfig {
@@ -6578,433 +6348,6 @@ impl TlonChannelConfig {
             self.resolved_account_identity().id.as_str(),
         )
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ResolvedConfiguredAccount {
-    id: String,
-    label: String,
-    account_key: Option<String>,
-}
-
-fn default_channel_account_identity() -> ChannelAccountIdentity {
-    ChannelAccountIdentity {
-        id: "default".to_owned(),
-        label: "default".to_owned(),
-        source: ChannelAccountIdentitySource::Default,
-    }
-}
-
-fn resolve_configured_account_identity(raw: Option<&str>) -> Option<(String, String)> {
-    let label = raw.map(str::trim).filter(|value| !value.is_empty())?;
-    if !label.chars().any(|value| value.is_ascii_alphanumeric()) {
-        return None;
-    }
-    Some((normalize_channel_account_id(label), label.to_owned()))
-}
-
-fn resolve_telegram_bot_id_from_token(token: &str) -> Option<&str> {
-    let bot_id = token.split(':').next()?.trim();
-    if bot_id.is_empty() || !bot_id.chars().all(|value| value.is_ascii_digit()) {
-        return None;
-    }
-    Some(bot_id)
-}
-
-fn resolve_string_with_legacy_env(raw: Option<&str>, env_key: Option<&str>) -> Option<String> {
-    let inline = raw
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned);
-    if inline.is_some() {
-        return inline;
-    }
-
-    let env_name = env_key.map(str::trim).filter(|value| !value.is_empty())?;
-    let env_value = std::env::var(env_name).ok()?;
-    let trimmed_value = env_value.trim();
-    if trimmed_value.is_empty() {
-        return None;
-    }
-    Some(trimmed_value.to_owned())
-}
-
-fn resolve_string_list_with_legacy_env(
-    raw: Option<&[String]>,
-    env_key: Option<&str>,
-) -> Vec<String> {
-    let inline = raw.map(normalize_inline_string_list).unwrap_or_default();
-    if !inline.is_empty() {
-        return inline;
-    }
-
-    let env_name = env_key.map(str::trim).filter(|value| !value.is_empty());
-    let Some(env_name) = env_name else {
-        return Vec::new();
-    };
-    let env_value = std::env::var(env_name).ok();
-    let Some(env_value) = env_value else {
-        return Vec::new();
-    };
-    parse_env_string_list(env_value.as_str())
-}
-
-fn normalize_inline_string_list(values: &[String]) -> Vec<String> {
-    values
-        .iter()
-        .map(String::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-        .collect()
-}
-
-fn parse_env_string_list(raw: &str) -> Vec<String> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return Vec::new();
-    }
-
-    trimmed
-        .split([',', '\n'])
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-        .collect()
-}
-
-pub(crate) fn parse_email_smtp_endpoint(raw: &str) -> CliResult<EmailSmtpEndpoint> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return Err("email smtp_host is empty".to_owned());
-    }
-
-    if trimmed.contains("://") {
-        let parsed_url = reqwest::Url::parse(trimmed)
-            .map_err(|error| format!("email smtp_host url is invalid: {error}"))?;
-        let scheme = parsed_url.scheme();
-        if scheme != "smtp" && scheme != "smtps" {
-            return Err(format!(
-                "email smtp_host url must use smtp:// or smtps://, got {scheme}://"
-            ));
-        }
-
-        let host = parsed_url
-            .host_str()
-            .map(str::trim)
-            .filter(|value| !value.is_empty());
-        if host.is_none() {
-            return Err("email smtp_host url is missing a host".to_owned());
-        }
-
-        return Ok(EmailSmtpEndpoint::ConnectionUrl(trimmed.to_owned()));
-    }
-
-    if trimmed.chars().any(char::is_whitespace) {
-        return Err("email smtp_host must not contain whitespace".to_owned());
-    }
-    if trimmed.contains('/') || trimmed.contains('?') || trimmed.contains('#') {
-        return Err(
-            "email smtp_host must be a bare host or a full smtp:// or smtps:// URL".to_owned(),
-        );
-    }
-    if trimmed.contains(':') {
-        return Err(
-            "email smtp_host with an explicit port must use a full smtp:// or smtps:// URL"
-                .to_owned(),
-        );
-    }
-
-    Ok(EmailSmtpEndpoint::RelayHost(trimmed.to_owned()))
-}
-
-pub(crate) fn normalize_channel_account_id(raw: &str) -> String {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return "default".to_owned();
-    }
-
-    let mut normalized = String::with_capacity(trimmed.len());
-    let mut last_was_separator = false;
-    for value in trimmed.chars() {
-        if value.is_ascii_alphanumeric() {
-            normalized.push(value.to_ascii_lowercase());
-            last_was_separator = false;
-            continue;
-        }
-        if matches!(value, '_' | '-') {
-            if !normalized.is_empty() && !last_was_separator {
-                normalized.push(value);
-                last_was_separator = true;
-            }
-            continue;
-        }
-        if !normalized.is_empty() && !last_was_separator {
-            normalized.push('-');
-            last_was_separator = true;
-        }
-    }
-
-    while matches!(normalized.chars().last(), Some('-' | '_')) {
-        normalized.pop();
-    }
-
-    if normalized.is_empty() {
-        "default".to_owned()
-    } else {
-        normalized
-    }
-}
-
-fn validate_email_env_pointer(
-    issues: &mut Vec<ConfigValidationIssue>,
-    field_path: &str,
-    env_key: Option<&str>,
-    inline_field_path: &str,
-) {
-    let example_env_name = if field_path.ends_with("imap_username_env") {
-        EMAIL_IMAP_USERNAME_ENV
-    } else if field_path.ends_with("imap_password_env") {
-        EMAIL_IMAP_PASSWORD_ENV
-    } else if field_path.ends_with("smtp_password_env") {
-        EMAIL_SMTP_PASSWORD_ENV
-    } else {
-        EMAIL_SMTP_USERNAME_ENV
-    };
-    if let Err(issue) = validate_env_pointer_field(
-        field_path,
-        env_key,
-        EnvPointerValidationHint {
-            inline_field_path,
-            example_env_name,
-            detect_telegram_token_shape: false,
-        },
-    ) {
-        issues.push(*issue);
-    }
-}
-
-fn validate_email_secret_ref_env_pointer(
-    issues: &mut Vec<ConfigValidationIssue>,
-    field_path: &str,
-    secret_ref: Option<&SecretRef>,
-) {
-    let example_env_name = if field_path.ends_with("imap_username") {
-        EMAIL_IMAP_USERNAME_ENV
-    } else if field_path.ends_with("imap_password") {
-        EMAIL_IMAP_PASSWORD_ENV
-    } else if field_path.ends_with("smtp_password") {
-        EMAIL_SMTP_PASSWORD_ENV
-    } else {
-        EMAIL_SMTP_USERNAME_ENV
-    };
-    if let Err(issue) = validate_secret_ref_env_pointer_field(
-        field_path,
-        secret_ref,
-        EnvPointerValidationHint {
-            inline_field_path: field_path,
-            example_env_name,
-            detect_telegram_token_shape: false,
-        },
-    ) {
-        issues.push(*issue);
-    }
-}
-
-fn configured_account_ids<'a, I>(keys: I) -> Vec<String>
-where
-    I: IntoIterator<Item = &'a String>,
-{
-    let mut ids = keys
-        .into_iter()
-        .map(|value| normalize_channel_account_id(value))
-        .collect::<Vec<_>>();
-    ids.sort();
-    ids.dedup();
-    ids
-}
-
-fn normalize_optional_account_id(raw: Option<&str>) -> Option<String> {
-    raw.map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(normalize_channel_account_id)
-}
-
-fn resolve_default_configured_account_selection_from_ids(
-    ids: &[String],
-    preferred: Option<&str>,
-    fallback: &str,
-) -> ChannelDefaultAccountSelection {
-    if let Some(preferred) = normalize_optional_account_id(preferred)
-        && !ids.is_empty()
-        && ids.iter().any(|value| value == &preferred)
-    {
-        return ChannelDefaultAccountSelection {
-            id: preferred,
-            source: ChannelDefaultAccountSelectionSource::ExplicitDefault,
-        };
-    }
-    if ids.is_empty() {
-        return ChannelDefaultAccountSelection {
-            id: normalize_channel_account_id(fallback),
-            source: ChannelDefaultAccountSelectionSource::RuntimeIdentity,
-        };
-    }
-    if ids.iter().any(|value| value == "default") {
-        return ChannelDefaultAccountSelection {
-            id: "default".to_owned(),
-            source: ChannelDefaultAccountSelectionSource::MappedDefault,
-        };
-    }
-    ChannelDefaultAccountSelection {
-        id: ids
-            .first()
-            .cloned()
-            .unwrap_or_else(|| normalize_channel_account_id(fallback)),
-        source: ChannelDefaultAccountSelectionSource::Fallback,
-    }
-}
-
-fn resolve_default_configured_account_selection<'a, I>(
-    keys: I,
-    preferred: Option<&str>,
-    fallback: &str,
-) -> ChannelDefaultAccountSelection
-where
-    I: IntoIterator<Item = &'a String>,
-{
-    let ids = configured_account_ids(keys);
-    resolve_default_configured_account_selection_from_ids(ids.as_slice(), preferred, fallback)
-}
-
-fn resolve_channel_account_route<'a, I>(
-    keys: I,
-    preferred: Option<&str>,
-    fallback: &str,
-    requested_account_id: Option<&str>,
-    selected_configured_account_id: &str,
-) -> ChannelResolvedAccountRoute
-where
-    I: IntoIterator<Item = &'a String>,
-{
-    let ids = configured_account_ids(keys);
-    let default_selection =
-        resolve_default_configured_account_selection_from_ids(ids.as_slice(), preferred, fallback);
-    ChannelResolvedAccountRoute {
-        requested_account_id: normalize_optional_account_id(requested_account_id),
-        configured_account_count: ids.len(),
-        selected_configured_account_id: normalize_channel_account_id(
-            selected_configured_account_id,
-        ),
-        default_account_source: default_selection.source,
-    }
-}
-
-fn resolve_channel_acp_config(
-    base: &ChannelAcpConfig,
-    account_override: Option<&ChannelAcpConfig>,
-) -> ChannelAcpConfig {
-    account_override.cloned().unwrap_or_else(|| base.clone())
-}
-
-fn resolve_account_for_session_account_id<R>(
-    session_account_id: Option<&str>,
-    resolve_direct: impl FnOnce() -> CliResult<R>,
-    configured_ids: impl FnOnce() -> Vec<String>,
-    resolve_configured: impl Fn(&str) -> CliResult<R>,
-    runtime_account_id: impl Fn(&R) -> &str,
-) -> CliResult<R> {
-    let Some(requested) = normalize_optional_account_id(session_account_id) else {
-        return resolve_direct();
-    };
-
-    match resolve_direct() {
-        Ok(resolved) => Ok(resolved),
-        Err(original_error) => {
-            for configured_id in configured_ids() {
-                let resolved = resolve_configured(configured_id.as_str())?;
-                if normalize_channel_account_id(runtime_account_id(&resolved)) == requested {
-                    return Ok(resolved);
-                }
-            }
-            Err(original_error)
-        }
-    }
-}
-
-fn resolve_configured_account_selection<'a, I>(
-    keys: I,
-    requested_account_id: Option<&str>,
-    preferred_default_account_id: Option<&str>,
-    fallback_id: &str,
-) -> CliResult<ResolvedConfiguredAccount>
-where
-    I: IntoIterator<Item = &'a String>,
-{
-    let entries = keys
-        .into_iter()
-        .filter_map(|value| {
-            let raw_key = value.to_owned();
-            let label = value.trim();
-            if label.is_empty() {
-                return None;
-            }
-            Some((
-                normalize_channel_account_id(label),
-                label.to_owned(),
-                raw_key,
-            ))
-        })
-        .collect::<Vec<_>>();
-    let configured_ids = entries
-        .iter()
-        .map(|(id, _, _)| id.clone())
-        .collect::<Vec<_>>();
-
-    if let Some(requested) = requested_account_id
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(normalize_channel_account_id)
-    {
-        if entries.is_empty() {
-            return Ok(ResolvedConfiguredAccount {
-                label: requested.clone(),
-                id: requested,
-                account_key: None,
-            });
-        }
-        let Some((id, label, raw_key)) = entries.iter().find(|(id, _, _)| *id == requested) else {
-            return Err(format!(
-                "requested account `{requested}` is not configured (configured accounts: {})",
-                configured_ids.join(", ")
-            ));
-        };
-        return Ok(ResolvedConfiguredAccount {
-            id: id.clone(),
-            label: label.clone(),
-            account_key: Some(raw_key.clone()),
-        });
-    }
-
-    let default_id = resolve_default_configured_account_selection(
-        entries.iter().map(|(_, _, raw_key)| raw_key),
-        preferred_default_account_id,
-        fallback_id,
-    )
-    .id;
-    if let Some((id, label, raw_key)) = entries.iter().find(|(id, _, _)| *id == default_id) {
-        return Ok(ResolvedConfiguredAccount {
-            id: id.clone(),
-            label: label.clone(),
-            account_key: Some(raw_key.clone()),
-        });
-    }
-
-    Ok(ResolvedConfiguredAccount {
-        id: default_id.clone(),
-        label: default_id,
-        account_key: None,
-    })
 }
 
 mod tlon_support;
