+++ crates/app/src/config/channels_defaults.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::prompt::{DEFAULT_PROMPT_PACK_ID, PromptPersonality, render_default_system_prompt};
 
 pub(super) fn default_telegram_base_url() -> String {
     "https://api.telegram.org".to_owned()
